requirements.txt
atlassian-python-api==3.26.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
idna==3.3
markdownify==0.11.2
oauthlib==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.11
wrapt==1.14.1
